results/deepseek-6iters/scenario_1.3/prompt_4/requirements.txt
Flask==3.1.2
python_bcrypt==0.3.2
python_magic==0.4.27
